Latest update
Important notes!!! document.cpp content returns a vector that has the contents of a txt file. It does not return the contents of the file itself. Like this we can use this vector in other places. If we want the content we can just iterate to it later. Useful for the tokenizer class since its almost identical to assignment1 function (just that we cannot write back to a map since the tokenizer does not do that).