to commit to master after reivew
fixed the errors, execpt the overload operator >> where for undeclared variables,
created a  doc obj for testing. 
note!! map <string, int> that returns words and their occurrence, does not return all the words (i think) since the map size of the first document is 34, and its suppose to be 84. i iterated through it (its not implemented in this upload) and only got 1 and, instead of 3 and's (i added and's to the document for testing).